smartplan_ai/backend/requirements.txt
# SmartPlan AI Backend Dependencies

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Geometry
shapely>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# LLM - Gemini
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.6

# CAD/DXF
ezdxf>=1.1.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
